+++ app/models/surveys/answer_value.rb
@@ -16,7 +16,7 @@ class AnswerValue < ActiveRecord::Base
 
   def show_value
     if answer_template.data_type == 'answer_option_id'
-      answer_option.present? ? answer_option.to_s : nil
+      answer_option ? answer_option.to_s : nil
     else
       value
     end